backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
Faker
passlib[bcrypt]
passlib[argon2]
argon2-cffi
python-jose[cryptography]
python-multipart
pytest 
pytest-asyncio
httpx